Revert "Merge branch 'master' of https://github.com/AlirezaNewTwen/ensmallen"
This reverts commit 62eeb615821b5059f0369debe74afc0e98ef4d7e, reversing
changes made to 5bb9fac6cf543ded1c69993a49b6fad285eef356.

+++ include/ensmallen_bits/parallel_sgd/decay_policies/exponential_backoff.hpp
@@ -64,7 +64,7 @@ class ExponentialBackoff
     if (numEpoch >= cutoffEpoch)
     {
       step *= beta;
-      cutoffEpoch += firstBackoffEpoch / static_cast<size_t>(beta);
+      cutoffEpoch += firstBackoffEpoch / beta;
     }
     return step;
   }
